+++ src/calc_genoprob_based_on_phased_marker_blocks.cpp
@@ -157,9 +157,9 @@ RcppExport SEXP calc_genprob_haplo(SEXP ploidyR,
       std::fill(gamma.begin(), gamma.end(), 0);
       int s = v[k][ind].size();
       long double w = 0.0;
-      for(int j=0; j < alpha[ind][k].size(); j++)
+      for(int j=0; (unsigned)j < alpha[ind][k].size(); j++)
         w += alpha[ind][k][j]*beta[ind][k][j];
-      for(int j=0; j < alpha[ind][k].size(); j++)
+      for(int j=0; (unsigned)j < alpha[ind][k].size(); j++)
         gamma[v[k][ind][j]*g + v[k][ind][j+s/2]] = alpha[ind][k][j]*beta[ind][k][j]/w;
       for(int j=0; j < g*g; j++)
       {
